kernel: Do not compile if UH, KDP and RKP enabled
UH, a kernel hypervisor driver that prevent any root permission.
If KernelSU driver was compiled with UH active, phone will freeze.

Need an accurate way to detect Samsung kernel, but currently we
depends on "drivers/samsung".

The reason why we don't put this on Kconfig is due to compatibility issue
with other OEMs.

Signed-off-by: rsuntk <rissu.ntk@gmail.com>